feat(keras, refactor): add initial files for Keras impl & test; refactor file names for PyTorch impl & test
